Include HKS for baseband modem M.2 module
Low-pass filter included for switch ckt, rise and fall time should be nearly identical with the presence of a schottky diode. Since the M.2 spec calls out a >20Kohm resistor to GND on the ~FULL_CARD_POWER_OFF pin then then a voltage divider is formed between this resistor and the one to VBAT (min of ~2.2V, actually >2.4V).