Add 1.8V<->3.3V logic level shifter (NTS0102) for the WWAN M.2 slot and connect it to the i.MX8M's UART3
The NTS0102 explicitly lists UART as one of its applications, and contains internal 10k pull-up resistors for the outputs. Whereas NTB/TXB0102 are push-pull outputs but aren't intended for UART applications due to the following (needs testing):
"The subtlety here is that you need to be careful of the auto direction sense circuitry.  Works fine but there can be some gotchas."
"If there is a signal integrity concern which is generally not the case for UARTs, series ~32 Ohm resistors on TX output pins can help.  For the exact resistor value you need to look at the V/I curve of the transmitter and pick the resistor value that matches the driver to the 50 Ohm Transmission line.  Layout for 50Ohm T-Lines and cables.  This will also cut down the risk of the EMI/EMC issues since any residual ringing is cutdown by IR drop in the terminator."