Replace instances of VSYS with VBAT to remove redundancy
In the case where VBAT can drop below 3.3V then it may be required to use a buck-boost in place of the existing buck converter for the 3.3V rail; may need some testing. Using the +5V rail as an input would also work but would introduce additional losses. According to several sources, when a 18650 Li-ion reaches 3.3V it is practically depleted anyway. When the feedback voltage drops below the internal reference voltage the buck converter will be in its "low dropout operation" with a duty cycle of 100%, so the output voltage should closely follow the input voltage.