Add I2C interface to WiFi/BT M.2, select an appropriate dual 2-input AND gate IC for the W_DISABLE logic and add the corresponding debounce/pull-down circuit to the HKS for W_DISABLE
Notes have been included regarding the I2C logic level voltage. If the USB interface is used then the input voltage needs to be 5.0V (not 3.3V); therefore the USB interface and hub may need to be removed!

Several signals are unused by the RS9116 WiFi/BT module, as noted in the schematic.