Add current sense resistors and jumper symbols in the schematic, assign the various signals left unspecified, complete the PWR_EN assignment and verify that thepower sequence would be brought up in the correct order, add an LED to the VBAT_REG rail for debugging purposes (know when the board is receiving power)
Both the GP_LED1 and the HAPTIC motor are able to be controlled via PWM2 by individually MUXing the corresponding pads to PWM2. Only one pin should be MUXed to PWM2 at a time whilst the other is a GPIO, otherwise they will behave identically.

The I2C buses do not need the bus switch found on the EmCraft baseboard since the M.2 modules' power is controlled by PWR_EN and their I2C is level shifted between 1.8V<->3.3V, making it safe to leave them as is.

The various sense resistors have their corresponding parts selected but will be "NC" in the BoM and will not be assembled during manufacturing. To measure the current on the various rails (and WLAN M.2), one will need to break the trace between the two pads and place the corresponding resistor. A jumper (1.27 or 2.45 still needs to be determined) can be used to re-short these pads.

Notes have been added regarding which GPIO pins can easily be interchanged if it improves the layout.

Selected a better/cheaper RJ45 receptacle. Also picked out a well-suited green LED to be used in various spots throughout the board.