Create footprint for cap touch FPC connector and assign it, add cap touch FPC connector 3D model, assign M.2 Key E socket footprint, assign JTAG header footprint (will be unpopulated)
M.2 footprints will need mounting holes added to them for the modules to be brought close to the baseboard, this may pose a problem with the display mounted on the back and might need to be done a different way (needs evaluation). Cap touch and DSI connectors need to be carefully placed in the layout to make sure they are positioned correctly with respect to the display's FPCs and the outline of the display itself (need dimensions ASAP). JTAG header can be placed on the back of the board with the header unpopulated, will be hidden as it's only intended for intense last resort debugging option.