Fit the new MHF4 footprint and antenna feedline together using the new footprint's dimensions
The MHF4 spec sheet gives dimensions for a 50Ohm CPW coming from it, however this feedline is actually a microstrip line and therefore it should follow our specific width requirements (using our relative permittivity and top-to-bottom spacing). The width of the feedline will need to be adjusted once the permittivity between the top and bottom layer with the inner layers being void is given/known (at ~3GHz).