Add stitching vias around the WiFi+BT antennae on the top-right section of the board, the vias form a fence-like structure and prevent interference in both directions
The rest of the path still needs to be stitched (around the GNSS antenna), and wrap around next to the earpiece speaker, which needed to be shifted slightly to the left. Once the proper via locations are determined between the speaker and the GNSS antenna area, the speaker can try to be placed as-close-as-possible to the horizontal-center of the board (at X=105mm).