Place and route the JTAG circuit, place and route the BOOT_MODE slide switch circuit, place and route the user-LED circuit, place and route the top-level sheet passive components (I2C2 pull-ups, 5V_SOM, 3V3_OUT, & ENET_2V5 bypass caps, PWR_EN pull-down, USB2_VBUS voltage divider for SoC's comparator, etc), route the microphone to the analog audio switch IC, update the schematic, BoM, and footprints to reflect the MHF4 coax cable change (including them being 5cm long)
Unrouted count is now 324. The power circuit and the display are what is now left to be placed on the board, a few nets currently on the board still need to be connected (~120). 3V3_SNVS and various other misc. power rails still need to be routed on the power plane (vias need to be placed near pads).

JTAG header footprint is accessible from the display side (not underneath the display). User-LED is in a location as-close-as-possible to the bottom center of the board (somewhat near the microphone).

Serpentine bends still need to be added to the CSI to match the intra-pair skew, still need to move the prox sensor to the top of the board!