Place and route the remaining 5V_P bypass caps which are spread throughout various spots on the board, get the unrouted count down to 0
Now that the unrouted count is 0, it is time to clean up the entire layout; improving various routing such as separating the SIM card slot's CLK & data and making the USB hub+SD chip's crystal routing symnetric. After that, the BoM needs to be updated with the passive components' tolerances, capacitors' temperature &  voltage ratings, and list recommended PNs which are shared with EmCraft's baseboard.