Modify the antennas' feedlines for a 1.1mm nominal thickness board using a relat ive permittivity of 4.1~4.11
The H parameter in the calculations was +0.21mm greater than the H used in the p
rototype calculations (1.1mm thick board instead of 0.89mm thick)

A DK value of 4.11 was used for the GNSS and WWAN antennae feedlines and a DK of 4.1 was used for the Wi-Fi/BT feedlines since the operating frequency of the GNSS and WWAN is generally a bit lower than Wi-Fi/BT.

The NANYA NP-180TL spec sheet is also included. The specific prepreg/core/substrate/dielectric used is NANYA NP-180**TL** which has a DK specified as 4.1-4.3 @ 1GHz.